tyrquake: widen Mod_FindInterval/Mod_GetSpriteFrame time arg to double
Both functions take 'float time' and are called as

  Mod_FindInterval(intervals, numframes, cl.time + e->syncbase);
  Mod_GetSpriteFrame(e, psprite, cl.time + e->syncbase);

cl.time is double in this fork, so the addition runs in double, but
the implicit narrowing at the call boundary truncates the result to
fp32. The internal math

  targettime = time - (int)(time / fullinterval) * fullinterval;

then runs at fp32 precision. ulp(fp32) at cl.time ~ 840,000 (~9.7 d
of continuous single-level engine time) is roughly 0.1, swamping the
per-frame animation intervals stored in the model data; at that point
animated alias-model skin groups, pose cycles, and animated sprites
start jittering or freezing because targettime quantizes to a coarser
grid than the interval table.

These functions are internal API (called only inside the engine, not
on any save/demo/network boundary), so just widen the time parameter
to double. The mod-and-truncate now runs in double; the final result
is bounded below fullinterval so the cast back to fp32 for the
comparison loop against the (small) interval table is exact.

R_AliasSetupFrame has its own copy of the mod math in the lerp path
('FIXME - merge with Mod_FindInterval?'); widen its local 'time' to
double the same way. The later 'time = targettime' / 'time = cl.time
- e->currentframetime' assignments still narrow back to small bounded
values, and the final blend = qclamp(time / delta, ...) keeps its
existing fp32 result type for the pose-blend factor.

Doesn't touch entity_t.previous*time / current*time fields (those
have the same antipattern but are part of a struct that ports may
treat as ABI-stable); those will need a separate countdown-style
conversion.

+++ common/model.h
@@ -528,8 +528,8 @@ void Mod_LoadAliasModel(const model_loader_t *loader, model_t *mod,
 void Mod_LoadSpriteModel(model_t *mod, void *buffer);
 
 mspriteframe_t *Mod_GetSpriteFrame(const struct entity_s *e,
-				   msprite_t *psprite, float time);
+				   msprite_t *psprite, double time);
 
-int Mod_FindInterval(const float *intervals, int numintervals, float time);
+int Mod_FindInterval(const float *intervals, int numintervals, double time);
 
 #endif /* MODEL_H */
